+++ CHANGELOG.md
@@ -16,7 +16,9 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Added a switch to enable/disable all application restrictions
+* Improved device registration procedure
 * Prevent pollution of usage data ([issue](https://github.com/M66B/XPrivacy/issues/1140))
+* Improved randomization of subscriber ID ([issue](https://github.com/M66B/XPrivacy/issues/1158))
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)