+++ CHANGELOG.md
@@ -15,12 +15,14 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 1.99.11 TEST**
+
 * Added a switch to enable/disable all application restrictions
 * Improved device registration procedure
 * Prevent pollution of usage data ([issue](https://github.com/M66B/XPrivacy/issues/1140))
 * Improved randomization of subscriber ID ([issue](https://github.com/M66B/XPrivacy/issues/1158))
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.10 TEST**